requirements.txt
anyio==4.1.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
dbus-python==1.2.16
Django==4.2.8
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
image==1.5.33
Pillow==10.1.0
pycups==1.9.73
PyGObject==3.36.0
python-telegram-bot==20.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==2.1.0
